Extract composer path tools
