+++ assets/js/editor-drafts.js
@@ -1,4 +1,4 @@
-import { readJsonStore, writeJsonStore } from './editor-storage.js?v=press-system-v3.4.30';
+import { readJsonStore, writeJsonStore } from './editor-storage.js?v=press-system-v3.4.31';
 
 export function createScopedDraftStore({
   storage,
