+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.30",
-  "tag": "v3.4.30",
+  "version": "3.4.31",
+  "tag": "v3.4.31",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.29 <3.4.30"
+      ">=3.4.30 <3.4.31"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.29 first."
+    "message": "Update to v3.4.30 first."
   }
 }
