+++ scripts/test-system-release-package.sh
@@ -106,6 +106,11 @@ if ! grep -qx "press-system-${version}/assets/js/composer-editor-shell.js" "${en
   exit 1
 fi
 
+if ! grep -qx "press-system-${version}/assets/js/composer-path-tools.js" "${entries_file}"; then
+  echo "expected package to include composer path tools code" >&2
+  exit 1
+fi
+
 if ! grep -qx "press-system-${version}/assets/js/editor-file-tree-ui.js" "${entries_file}"; then
   echo "expected package to include editor file tree UI code" >&2
   exit 1
